backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
PyYAML==6.0.1
python-multipart==0.0.6
faiss-cpu==1.7.4
transformers==4.35.0
torch==2.1.0
tqdm==4.66.1
numpy==1.24.3 